+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -137,7 +137,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-01T06:04:38.639Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-01T11:45:49.928Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")