+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -1107,7 +1107,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1DS3XQ0Z98D7KDKME4KKSM6"
+        "tool_01M1ECMMXK6TG1T2FR0W7NQ4JN"
     )
 
     display_name: Optional[str] = None
@@ -1412,7 +1412,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1DS3XPZYXNYYMPFY5VE5F7J"
+        "tool_01M1ECMMXJXBJRDBQHNRRVDBZG"
     )
 
     display_name: Optional[str] = None
@@ -1606,7 +1606,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1DS3XPYS70B7VA7WPWAB5SG"
+        "tool_01M1ECMMXJCXMYFWH2152F1PC6"
     )
 
     display_name: Optional[str] = None
@@ -1804,7 +1804,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1DS3XPX72M8B9TMB82T2XSJ"
+        "tool_01M1ECMMXHWVCB7SYWXV9X0VH9"
     )
 
     display_name: Optional[str] = None