+++ packages/orq-rc/RELEASES.md
@@ -10579,3 +10579,13 @@ Based on:
 - [python v4.15.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.9 - packages/orq-rc
+
+## 2026-09-01 11:48:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.2 (2.933.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.11] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.11 - packages/orq-rc